requirements_command_center.txt
# Trinity Command Center Additional Dependencies

# Web Framework (Streamlit)
streamlit>=1.30.0

# 3D/CAD Processing
trimesh>=4.0.0
numpy-stl>=3.0.0

# System Monitoring
psutil>=5.9.0

# Already required by Trinity (included for completeness)
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Optional: For advanced 3D processing
# pymeshlab>=2022.2  # Uncomment if you want mesh optimization
